+++ vault/resource_ldap_auth_backend.go
@@ -3,26 +3,26 @@ package vault
 import (
 	"fmt"
 	"log"
-	"regexp"
 	"strings"
 
 	"github.com/hashicorp/terraform/helper/schema"
+
 	"github.com/hashicorp/vault/api"
 )
 
-var (
-	ldapAuthBackendConfigFromPathRegex = regexp.MustCompile("^auth/(.+)/config$")
-)
+const ldapAuthType string = "ldap"
 
-func ldapAuthBackendConfigResource() *schema.Resource {
+func ldapAuthBackendResource() *schema.Resource {
 	return &schema.Resource{
 		SchemaVersion: 1,
 
-		Create: ldapAuthBackendConfigCreate,
-		Update: ldapAuthBackendConfigUpdate,
-		Read:   ldapAuthBackendConfigRead,
-		Delete: ldapAuthBackendConfigDelete,
-		Exists: ldapAuthBackendConfigExists,
+		Create: ldapAuthBackendWrite,
+		Update: ldapAuthBackendUpdate,
+		Read:   ldapAuthBackendRead,
+		Delete: ldapAuthBackendDelete,
+		Exists: ldapAuthBackendExists,
+
+		DeprecationMessage: "ldap_auth_backend resource is deprecated and will be removed with version 1.5 - use ldap_auth_backend_config",
 
 		Schema: map[string]*schema.Schema{
 			"url": {
@@ -108,42 +108,59 @@ func ldapAuthBackendConfigResource() *schema.Resource {
 				Optional: true,
 				Computed: true,
 			},
-			"backend": {
-				Type:        schema.TypeString,
-				Optional:    true,
-				Description: "Unique name of the ldap backend to configure.",
-				ForceNew:    true,
-				Default:     "ldap",
-				// standardise on no beginning or trailing slashes
+
+			"description": {
+				Type:     schema.TypeString,
+				Optional: true,
+				Computed: true,
+			},
+
+			"path": {
+				Type:     schema.TypeString,
+				Optional: true,
+				ForceNew: true,
+				Default:  "ldap",
 				StateFunc: func(v interface{}) string {
 					return strings.Trim(v.(string), "/")
 				},
+			},
+
+			"accessor": {
+				Type:        schema.TypeString,
+				Computed:    true,
+				Description: "The accessor of the LDAP auth backend",
 			},
 		},
 	}
 }
 
-func baseAuthBackendConfigPath(path string) string {
+func ldapAuthBackendConfigPath(path string) string {
 	return "auth/" + strings.Trim(path, "/") + "/config"
 }
 
-func ldapAuthBackendConfigBackendFromPath(path string) (string, error) {
-	if !ldapAuthBackendConfigFromPathRegex.MatchString(path) {
-		return "", fmt.Errorf("no backend found")
+func ldapAuthBackendWrite(d *schema.ResourceData, meta interface{}) error {
+	client := meta.(*api.Client)
+
+	authType := ldapAuthType
+	path := d.Get("path").(string)
+	desc := d.Get("description").(string)
+
+	log.Printf("[DEBUG] Enabling LDAP auth backend %q", path)
+	err := client.Sys().EnableAuth(path, authType, desc)
+	if err != nil {
+		return fmt.Errorf("error enabling ldap auth backend %q: %s", path, err)
 	}
-	res := ldapAuthBackendConfigFromPathRegex.FindStringSubmatch(path)
-	if len(res) != 2 {
-		return "", fmt.Errorf("unexpected number of matches (%d) for backend", len(res))
-	}
-	return res[1], nil
+	log.Printf("[DEBUG] Enabled LDAP auth backend %q", path)
+
+	d.SetId(path)
+
+	return ldapAuthBackendUpdate(d, meta)
 }
 
-func ldapAuthBackendConfigCreate(d *schema.ResourceData, meta interface{}) error {
+func ldapAuthBackendUpdate(d *schema.ResourceData, meta interface{}) error {
 	client := meta.(*api.Client)
-	backend := d.Get("backend").(string)
-	path := baseAuthBackendConfigPath(backend)
-	log.Printf("[DEBUG] Writing ldap auth backend config %q", path)
 
+	path := ldapAuthBackendConfigPath(d.Id())
 	data := map[string]interface{}{}
 
 	if v, ok := d.GetOk("url"); ok {
@@ -212,8 +229,6 @@ func ldapAuthBackendConfigCreate(d *schema.ResourceData, meta interface{}) error
 
 	log.Printf("[DEBUG] Writing LDAP config %q", path)
 	_, err := client.Logical().Write(path, data)
-
-	d.SetId(path)
 
 	if err != nil {
 		d.SetId("")
@@ -221,117 +236,41 @@ func ldapAuthBackendConfigCreate(d *schema.ResourceData, meta interface{}) error
 	}
 	log.Printf("[DEBUG] Wrote LDAP config %q", path)
 
-	return ldapAuthBackendConfigRead(d, meta)
+	return ldapAuthBackendRead(d, meta)
 }
 
-func ldapAuthBackendConfigUpdate(d *schema.ResourceData, meta interface{}) error {
+func ldapAuthBackendRead(d *schema.ResourceData, meta interface{}) error {
 	client := meta.(*api.Client)
+
 	path := d.Id()
-
-	log.Printf("[DEBUG] Updating ldap auth backend config %q", path)
-
-	data := map[string]interface{}{}
-
-	if v, ok := d.GetOk("url"); ok {
-		data["url"] = v.(string)
-	}
-
-	if v, ok := d.GetOkExists("starttls"); ok {
-		data["starttls"] = v.(bool)
-	}
-
-	if v, ok := d.GetOk("tls_min_version"); ok {
-		data["tls_min_version"] = v.(string)
-	}
-
-	if v, ok := d.GetOk("tls_max_version"); ok {
-		data["tls_max_version"] = v.(string)
-	}
-
-	if v, ok := d.GetOkExists("insecure_tls"); ok {
-		data["insecure_tls"] = v.(bool)
-	}
-
-	if v, ok := d.GetOk("certificate"); ok {
-		data["certificate"] = v.(string)
-	}
-
-	if v, ok := d.GetOk("binddn"); ok {
-		data["binddn"] = v.(string)
-	}
-
-	if v, ok := d.GetOk("bindpass"); ok {
-		data["bindpass"] = v.(string)
-	}
-
-	if v, ok := d.GetOk("userdn"); ok {
-		data["userdn"] = v.(string)
-	}
-
-	if v, ok := d.GetOk("userattr"); ok {
-		data["userattr"] = v.(string)
-	}
-
-	if v, ok := d.GetOkExists("discoverdn"); ok {
-		data["discoverdn"] = v.(bool)
-	}
-
-	if v, ok := d.GetOkExists("deny_null_bind"); ok {
-		data["deny_null_bind"] = v.(bool)
-	}
-
-	if v, ok := d.GetOk("upndomain"); ok {
-		data["upndomain"] = v.(string)
-	}
-
-	if v, ok := d.GetOk("groupfilter"); ok {
-		data["groupfilter"] = v.(string)
-	}
-
-	if v, ok := d.GetOk("groupdn"); ok {
-		data["groupdn"] = v.(string)
-	}
-
-	if v, ok := d.GetOk("groupattr"); ok {
-		data["groupattr"] = v.(string)
-	}
-
-	log.Printf("[DEBUG] Writing LDAP config %q", path)
-	_, err := client.Logical().Write(path, data)
+	auths, err := client.Sys().ListAuth()
 	if err != nil {
-		return fmt.Errorf("error updating ldap config %q: %s", path, err)
+		return fmt.Errorf("error reading from Vault: %s", err)
 	}
 
-	// NOTE: Only `SetId` after it's successfully written in Vault
-	d.SetId(path)
-
-	log.Printf("[DEBUG] Wrote LDAP config %q", path)
-
-	return ldapAuthBackendConfigRead(d, meta)
-}
-
-func ldapAuthBackendConfigRead(d *schema.ResourceData, meta interface{}) error {
-	client := meta.(*api.Client)
-	path := d.Id()
-
-	backend, err := ldapAuthBackendConfigBackendFromPath(path)
-	if err != nil {
-		return fmt.Errorf("invalid path %q for ldap auth backend config: %s", path, err)
+	authMount := auths[strings.Trim(path, "/")+"/"]
+	if authMount == nil {
+		return fmt.Errorf("auth mount %s not present", path)
 	}
 
-	log.Printf("[DEBUG] Reading ldap auth backend config %q", path)
+	d.Set("description", authMount.Description)
+	d.Set("accessor", authMount.Accessor)
+
+	path = ldapAuthBackendConfigPath(path)
+
+	log.Printf("[DEBUG] Reading LDAP auth backend config %q", path)
 	resp, err := client.Logical().Read(path)
 	if err != nil {
 		return fmt.Errorf("error reading ldap auth backend config %q: %s", path, err)
 	}
-	log.Printf("[DEBUG] Read ldap auth backend config %q", path)
+	log.Printf("[DEBUG] Read LDAP auth backend config %q", path)
+
 	if resp == nil {
-		log.Printf("[WARN] ldap auth backend config %q not found, removing from state", path)
+		log.Printf("[WARN] LDAP auth backend config %q not found, removing from state", path)
 		d.SetId("")
 		return nil
 	}
 
-	d.Set("backend", backend)
 	d.Set("url", resp.Data["url"])
 	d.Set("starttls", resp.Data["starttls"])
 	d.Set("tls_min_version", resp.Data["tls_min_version"])
@@ -354,24 +293,30 @@ func ldapAuthBackendConfigRead(d *schema.ResourceData, meta interface{}) error {
 	return nil
 }
 
-func ldapAuthBackendConfigDelete(d *schema.ResourceData, meta interface{}) error {
+func ldapAuthBackendDelete(d *schema.ResourceData, meta interface{}) error {
+	client := meta.(*api.Client)
 	path := d.Id()
-	log.Printf("[DEBUG] Deleted ldap auth backend config %q", path)
-	d.SetId("")
+
+	log.Printf("[DEBUG] Deleting LDAP auth backend %q", path)
+	err := client.Sys().DisableAuth(path)
+	if err != nil {
+		return fmt.Errorf("error deleting ldap auth backend %q: %q", path, err)
+	}
+	log.Printf("[DEBUG] Deleted LDAP auth backend %q", path)
+
 	return nil
 }
 
-func ldapAuthBackendConfigExists(d *schema.ResourceData, meta interface{}) (bool, error) {
+func ldapAuthBackendExists(d *schema.ResourceData, meta interface{}) (bool, error) {
 	client := meta.(*api.Client)
+	path := ldapAuthBackendConfigPath(d.Id())
 
-	path := d.Id()
-	log.Printf("[DEBUG] Checking if ldap auth backend config %q exists", path)
-
+	log.Printf("[DEBUG] Checking if LDAP auth backend %q exists", path)
 	resp, err := client.Logical().Read(path)
 	if err != nil {
-		return true, fmt.Errorf("error checking if ldap auth backend config %q exists: %s", path, err)
+		return true, fmt.Errorf("error checking for existence of ldap config %q: %s", path, err)
 	}
-	log.Printf("[DEBUG] Checked if ldap auth backend config %q exists", path)
+	log.Printf("[DEBUG] Checked if LDAP auth backend %q exists", path)
 
 	return resp != nil, nil
 }